atualização para dotnet 6

+++ src/FinoSabor.Domain/Entities/Itens_Pedido.cs
@@ -1,16 +1,12 @@
 ﻿using FinoSabor.Domain.Entities.Base;
 using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations.Schema;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace FinoSabor.Domain.Entities
 {
     public class Itens_Pedido : EntityBase
     {
-        public int valor_unitario { get; set; }
+        public decimal valor_unitario { get; set; }
         public int quantidade { get; set; }
 
         [NotMapped] public decimal valor_item { get; set; }
